# [Backport to 5.11] Pre-compiled regexp optimization