Baseline COSBENCH

Op-Type	Op-Count	Byte-Count	Avg-ResTime	Avg-ProcTime	Throughput	Bandwidth	Succ-Ratio op1: read	294.63 kops	2.95 GB	232.11 ms	232.11 ms	327.42 op/s	3.27 MB/S	100%

PR COSBENCH

Op-Type	Op-Count	Byte-Count	Avg-ResTime	Avg-ProcTime	Throughput	Bandwidth	Succ-Ratio op1: read	339.79 kops	3.4 GB	201.28 ms	201.28 ms	377.58 op/s	3.78 MB/S	100%

Signed-off-by: Alexander Indenbaum <aindenba@redhat.com>
(cherry picked from commit c3393dc38a97dbc8e9de97c3e0f299dbc918a38d)

### Explain the changes
1. 

### Issues: Fixed #xxx / Gap #xxx
1. 

### Testing Instructions:
1. 


- [ ] Doc added/updated
- [ ] Tests added